Tracking collected pickups
Collected pickups are shown in the top right.
Changed Level 1 to have four pickups, one at each corner.
Refactored drawing of pickups and movers so that x,y are coordinates for 8x8 sprite. It's now up to the map to adjust this as needed. This facilitates drawing of items outside of the map.