Fixed drawing of movers
Movers could be clipped by the unit towards the camera that they were moving from or are moving to. This is fixed by moving it front unit as soon as it is on top of it. This is distinct from the when it enters the unit for game play purposes (pick ups, and player death).

Still pending: Tree-based iso-lines. The player can be incorrectly drawn in front of some neighbouring tiles on the iso-line. This is very obvious with some permanent high tiles and early movement to the front-most tile.