Improved handling of big height differences
Limiting wave height by applying a smooth_clamp function.
Made drawing of bottom part of sprite conditional on current height to ensure this is only done when needed, and tile sprites are not drawn below the borden when the height is too low.
Also increased height of border sprites to ensure nothing leaks through below them. Rearranged sprites slightly for this.