Added support for larger levels
Created basic 16x16 Level 4.
Added camera movement. When close to the borders of the map, the camera does not try to center the player. This avoids too much empty space. Also, it ensures that there's no scrolling for the smaller 8x8 maps.
Also some work in scratch sprite area to see if display of checkered tiles can be improved. This is still to be continued.
Note: This commit also contains changes that should have been part of f1cb148e