Minor improvements to tile sprites
For tiles there is a slight conflict. For a perfect fitting square top-surface, the tile is only 15-pixels wide. This means that at one side of the map, one pixel is still showing for the side-surface that should be fully obscured, which does not look good. Making the tile 16-pixels wide fixes this, but gives a top-surface that is one-pixel too wide, which can also look ugly. This is resolved by using a 16-pixel wide tile for static square tiles which can border the map in both directions, and a 15-pixel wide tiles otherwise.