Improved drawing of Movers
For movers at same map unit, start drawing the one that is furthest back.
Updated enemy move logic so that it does not move to unit with another mover except the target. This should prevent units with more than two movers (which avoids having to draw this properly) and avoids two enemies following the same path.