Improved box interactions
Do allow box to be dropped onto another box, which will destroy it.
Do not allow box to be pushed onto a pick-up.
A mover falling on a box will only destroy a box that is not moving. This to avoid a push from slightly higher trigger a box destroy sound when the player enters the unit of the box.