Improved drop animation of box
The old drop animation did not work for tiles where part of its side was showing, which was the case for new "Beam me Up" level. Therefore resorted to old idea to use palette to only partially draw box. Removed support for draw2 method, as it is not needed anymore.