Improved drop animation
Added spr_dropped function for drawing a mover that may be dropping. If it is, it dynamically creates a new sprite, obscure part of the source sprite based on the amount it drops.
Also let player become dazed after it dropped. Also delayed signalling death so that drop sound effect is fully finished.
Moved update of height when dropping from update_dx_dy method to update_height method for both Player and Box.
Added test level with some gaps and boxes to test the above.