Fix to isoline drawing code
Movers could still be drawn on top of tiles that should obscure them. This has been fixed. In doing so, refactored the underlying model. The height method now takes a boolean to indicate if the left- or right-side is desired. This is needed as for one given isoline segment the height at both ends may be needed, in different contexts.