Added watermark to end screen
The watermark is overlayed on the bottom text. It is a binary encoding of all the best level scores. This can be used as proof of the actual solve. Different people should have a different watermark. The scores should match the virtual high-score. Finally, the scores need to be realistic. If they are too low, then it signals that a level was not completed. If they are too high, it also signals a cheat.