Bug fix: Do not collect pickup once dying
It used to be possible to finish a level after death was signalled. This could for example occur in level "Aquatic Race", where time-out was signaled when the player was entering the unit with the final pickup. In this case, even though the player was frozen, the pick-up would still be collected.