Tracking level run (and max level run)
Cartdata now also stores the maximum run of levels that has been completed in a game.
Updated end screen to show the level run and the maximum level run.
When reaching the end of the line, the "Game Done" animation is now skipped as it had become a stripped down end screen. The "Game Done" animation has been renamed to "Game Over" to reflect its new role.