Minor fix to avoid rendering artifact
When about to enter a unit closer to the camera, some pixels of the bot were obscured by this new unit. This is fixed by changing the rounding of the y-delta. An additional benefit is that the change of dx and dy happens at slightly different moments, ensuring a more gradual movement.