Smoother bot movement
Improved bot movement so that changes are as gradual as possible. In particular, the changes of x and y coordinate now always happen in different frames. This uses same logic as used in Gamebuino implementation.
While doing so, hard-coded movement delay to 2. Otherwise, logic has to be more complex. Same as was done with Gamebuino implementation.

Also updated camera-movement. It is now such that player remains exactly at center of screen, without moving a pixel up or down, or left or right. This is much easier on the eye. Also here logic is now the same as for Gamebuino.